really working UMD version
