+++ package.json
@@ -22,9 +22,9 @@
     "@webreflection/signal": "^2.0.0",
     "es-module-shims": "^1.6.2",
     "esbuild": "^0.15.14",
-    "esbuild-plugin-umd-wrapper": "^2.0.0",
     "hyperf": "^1.6.2",
     "jsdom": "^21.1.0",
+    "rollup": "^4.18.1",
     "signal-polyfill": "^0.1.1",
     "subscript": "^8.3.5",
     "tst": "^7.1.1",
@@ -33,7 +33,7 @@
   },
   "scripts": {
     "test": "node -r ./test/register.cjs test/test.js",
-    "build": "node .esbuild.js",
+    "build": "node .build.js",
     "watch": "esbuild --bundle --watch ./sprae.js --outdir=dist --format=esm"
   },
   "repository": {
