+++ sprae.init.js
@@ -1,0 +1,5 @@
+import sprae from './sprae.js'
+
+if (document?.currentScript?.hasAttribute('init')) sprae(document.documentElement)
+
+export default sprae
